feat: update the `api.mustache` template and go generator to support JSON media type for the request body
The Go Generator has been updated to add a vendor extension to set `x-is-json` to true when the consume media type is `application/json` and `x-is-form` when the consumes media type is `application/x-www-form-urlencoded`

The `x-is-json` vendor extension is used to generate the code to handle calling the new PostJson function or the corresponding HTTP method function on the request handler. This is designed to only work with Post requests at the moment but could be extended in the future if needed.

This PR aids in resolving https://github.com/twilio/twilio-oai-generator/issues/49 however once this PR is merged https://github.com/twilio/twilio-oai/issues/36 will need to be finished to fully resolve the issue

This change relies on https://github.com/twilio/twilio-go/pull/83 and the tests cannot be updated until this PR is merged and released

This change also fixes an issue with JSON struct tags for the Params structs being `html-escaped`. I have disabled the escaping by using `{{{}}}` as this was highlighted during linting the Go SDK repo