Fix compile error introduced by UpdateAlarmStatus
As part of the introduction of `UpdateAlarmStatus` to twilio-oai, there
is now an API method that calls an update API endpoint that does not
return any content. This seems to be the first instance of an API
endpoint that falls into this category, and it has introduced an issue
with the partial serialization template: the partial serialization
template expected the function that the template is being embedded in
always has a return value.

This commit fixes the issue by checking to see if there is an expected
return type and if not switches to returning only an error.